feat : useChangePage hook #39
