+++ src/components/Home/Main/index.tsx
@@ -2,30 +2,18 @@ import * as S from "./style";
 import { DodamSegmentedButton } from "@b1nd/dds-web";
 import PersonalNightStudy from "components/Common/PersonalNightStudy";
 import ProjcetNightStudy from "components/Common/ProjectNightStudy";
-import { SetStateAction, useState } from "react";
+import { useChangePage } from "hooks/NightStudy/useChangePage";
 
 interface Props {
   isPersonalPage: boolean;
-  setIsPersonalPage: React.Dispatch<SetStateAction<boolean>>
-}
-
-interface PageDataType {
-  text: string;
-  isAtv: boolean;
+  setIsPersonalPage: React.Dispatch<React.SetStateAction<boolean>>;
 }
 
 const Main = ({ isPersonalPage, setIsPersonalPage }: Props) => {
-  const [pageData, setPageData] = useState<PageDataType[]>([
+  const { handleClickPage, pageData } = useChangePage([
     { text: '개인', isAtv: isPersonalPage },
     { text: '프로젝트', isAtv: !isPersonalPage }
-  ])
-
-  const handleClickPage = (text?: string) => {
-    setPageData((prev) =>
-      prev.map((item) => ({ ...item, isAtv: item.text === text }))
-    );
-    setIsPersonalPage(text === "개인" ? true : false);
-  };
+  ], setIsPersonalPage)
   
   return (
     <S.Container>
