+++ src/components/Home/Sidebar/index.tsx
@@ -1,26 +1,16 @@
 import * as S from "./style";
-import { Suspense, useState } from "react";
+import { Suspense } from "react";
 import { DodamErrorBoundary, DodamSegmentedButton } from "@b1nd/dds-web";
 import MyNightStudy from "components/Common/MyNightStudy";
 import MyNightStudyFallback from "components/Common/Fallback/MyNightStudyFallback";
 import IsBannedChecker from "components/Common/IsBannedChecker";
-
-interface PageDataType {
-  text: string;
-  isAtv: boolean;
-}
+import { useChangePage } from "hooks/NightStudy/useChangePage";
 
 const Sidebar = ({ isPersonalPage }: { isPersonalPage: boolean }) => {
-  const [pageData, setPageData] = useState<PageDataType[]>([
+  const { handleClickPage, pageData } = useChangePage([
     { text: "대기중", isAtv: true },
     { text: "승인됨", isAtv: false },
-  ]);
-
-  const handleClickPage = (text?: string) => {
-    setPageData((prev) =>
-      prev.map((item) => ({ ...item, isAtv: item.text === text }))
-    );
-  };
+  ])
 
   return (
     <S.SideBarContainer>
